refactor templated SARSALearner into its own header
