+++ src/sarsa_agent.h
@@ -33,105 +33,6 @@ class Experience {
   virtual double PredictScore() const = 0;
 };
 
-template <int N>
-class SARSALearner;
-
-template <int N>
-class ExperienceImpl : public Experience {
- public:
-  ExperienceImpl(std::unique_ptr<Action> action, double score,
-                 Experience* next_experience, double features[N],
-                 SARSALearner<N>* learner)
-      : Experience(std::move(action), score, next_experience),
-        learner_(learner) {
-          for(int i=0; i<N; i++) {
-            features_[i] = features[i];
-          }
-        }
-
-  double features_[N];
-  SARSALearner<N>* learner_;
-
-  double Learn(CVC* cvc) override;
-
-  double PredictScore() const override {
-    return learner_->Score(features_);
-  }
-};
-
-template <int N>
-class SARSALearner {
- public:
-  static void ReadWeights(
-      const char* weight_file,
-      std::unordered_map<std::string, SARSALearner*> learners);
-
-  static void WriteWeights(
-      const char* weight_file,
-      std::unordered_map<std::string, SARSALearner*> learners);
-
-
-  //creates a randomly initialized learner
-  static std::unique_ptr<SARSALearner> Create(int learner_id, double n,
-                                              double g, double b1, double b2,
-                                              std::mt19937& random_generator,
-                                              Logger* learn_logger);
-
-  SARSALearner(int learner_id, double n, double g, double b1, double b2,
-               double weights[N], Stats s[N],
-               double m[N],  double r[N],
-               Logger* learn_logger)
-    : learner_id_(learner_id),
-      n_(n),
-      g_(g),
-      b1_(b1),
-      b2_(b2),
-      learn_logger_(learn_logger) {
-    for(int i=0; i<N; i++) {
-      weights_[i] = weights[i];
-      feature_stats_[i] = s[i];
-      m_[i] = m[i];
-      r_[i] = r[i];
-    }
-  }
-
-  double Learn(CVC* cvc, ExperienceImpl<N>* experience);
-
-  void WriteWeights(FILE* weights_file);
-  void ReadWeights(FILE* weights_file);
-
-  double Score(const double features[N]) const;
-  double ComputeDiscountedRewards(const Experience* experience) const;
-
-  std::unique_ptr<Experience> WrapAction(double features[N],
-                                         std::unique_ptr<Action> action) {
-    action->SetScore(Score(features));
-    return std::make_unique<ExperienceImpl<N>>(std::move(action), 0.0, nullptr,
-                                               features, this);
-  }
-
- private:
-  int learner_id_;
-
-  double n_; //learning rate
-  double g_; //discount factor
-  double weights_[N];
-
-  Stats feature_stats_[N];
-
-  //adam optimizer params and state
-  double b1_;
-  double b2_;
-  //TODO: parametrize ADAM epsilon?
-  double epsilon_ = .000000001; //10^-8
-  //TODO: setting learning epoch always to 0 means we can't load state
-  //so this would need to get serialized out
-  int t_=0;
-  double m_[N];
-  double r_[N];
-
-  Logger* learn_logger_;
-};
 
 // just one kind of action, one model
 class SARSAActionFactory {
